refactor: migrate QuackText V1 to V2
