+++ feature/setting/src/main/kotlin/team/duckie/app/android/feature/setting/screen/SettingInquiryScreen.kt
@@ -22,7 +22,8 @@ import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import team.duckie.app.android.feature.setting.R
 import team.duckie.app.android.feature.setting.component.SettingContentLayout
-import team.duckie.quackquack.ui.component.QuackSubtitle2
+import team.duckie.quackquack.material.QuackTypography
+import team.duckie.quackquack.ui.QuackText
 
 /**
  * 덕키 공식 이메일
@@ -46,9 +47,10 @@ fun SettingInquiryScreen() {
     Column(
         modifier = Modifier.fillMaxSize(),
     ) {
-        QuackSubtitle2(
+        QuackText(
             modifier = Modifier.padding(vertical = 12.dp),
             text = stringResource(id = R.string.contact_inquiry),
+            typography = QuackTypography.Subtitle2,
         )
         LazyColumn(
             verticalArrangement = Arrangement.spacedBy(4.dp),
