+++ feature/setting/src/main/kotlin/team/duckie/app/android/feature/setting/screen/SettingMainPolicyScreen.kt
@@ -16,10 +16,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import team.duckie.app.android.common.compose.ui.Spacer
 import team.duckie.app.android.feature.setting.R
 import team.duckie.app.android.feature.setting.component.SettingContentLayout
 import team.duckie.app.android.feature.setting.constans.SettingType
-import team.duckie.app.android.common.compose.ui.Spacer
 
 @Composable
 fun SettingMainPolicyScreen(
